fix(frontend): Refactor global state and dispatches
I've refactored the global state to pass `filterRange` to the slider component as its `brushDomain`, and updated the dispatches/useEffect to only trigger when and what it should, so that the component performs as it should.

This includes only dispatching the property's min/max ranges is there are none in state, and populating the slider with the selected property's ranges is they are present. On every click of a toggle or button, the clipPath is reset.