fix(frontend): Update reducers, add property as prop
I updated the `IsolationPlus` reducers so that, instead of updating just the filterRange, they take an object comprised of the filterRange and the clipPosition. This is because the component needs to have knowledge of the clip position in order to reset the local state upon property change, but the values for clipPosition come from the brysh's chane handler.

To hydrate the component'sclipPosition with the currently selected position, it needs to come from global state, and this is the best way I could see to do that.

Also, I passed the currently selected property down to the component as a prop, so that it can be used in the `useEffect` to reset the local state. The reason the `filterRange` or `clipPath` cannot be changed is because the componsnt cannot tell the difference between the state a) changing from moving the slider, or b) changing from selecting a new property.