fix(frontend): Refactor sidebar and map components
Refactored sidebar and map components to be re-used across multiple layers of the same type.

It is all hooked up as intended, and (sort of) works.
The issue at the moment is that the unique identifying properties for the sidebar (source_id) and feature detail (name prop on map_component), are given to all layers simultaneously.

This method works in theory, but I believe to be caused by all of the sources being mapped in `useOrbs.js`, returning all before the iteration ends.