fix(frontend): Revert AFH files, modify pinIconConfig
Reverted the AFH files to their original state, as now using the `pinIconConfig`, but wrapping in a higher level for the `MSL` data, so that the categories can be filtered before it hits the config.

This ensures that the same structure of data is hitting the config every time, meaning that all of the pin-style layers can now use it.