fix(frontend): Switched to using an array of 1, simplified component.
Instead of returning single properties as single objects, the helper fruntion now puts single properties inside an array also, so not the `RadioProperty` is always receiving an array.

This allows the component's logic to be greatly simplified.

Also switched buttons around, so that `Number` is first, as it seems to always be the default.