fix(frontend): Initial code to solve issue.
The code here solves the issue by doing what is asked.
There is room for refactoring, tests, typing, etc, but for now it's just to test faesability.

Changes are basically a multi-level builder to construct separate state objects for all of the changed sliders.
They will persist across properties, layers and orbs, for the duration of the session.

It would be a good idea to clear the state when the layer is removed from the map, but one thing at a time.