fix(frontend): Add calculations to top-level, clean up styles
Moved the calculations of the analysis panel to the top-level, and dispatched along with the screenshot.

Althought the eventual move would be to have these calculations be the only ones, for now, they are still in the individual components until this solution has been confirmed.

Also, cleaned up styles to match designs.