fix(frontend): Add button banner to PDF document
Before, there was a single utton to download the document, which was absolutely positioned.

Now, with the addition of a seconf button (close icon), I have changed it to be the 2 buttons inside a container, which is absolutely positioned on top of the PDF component.

Also, passed the close function to the component to explicitly close it when the button is clicked, and updated tests stories.