fix(frontend): Add logging of whether property toggled on/off
It makes sense to log the state, rather than just whether the property
has been toggled. Needed to move the code to below where the state
is being set, so it correctly reflects what has happened.

There was code to do this check, I've just moved it up to be it's own
function and called it from the 2 places it is referenced.

IssueID ORB-223