fix(frontend): Prohibit area highlighting unless from pill, add styles
The structure of state-controlled coloring was caausing the area to highlight white even when hovered from the are itself, not the pill.

When the area is hovered, the autoHighlight should be used, not the global state check.
In order to acheive this, a `type` property was added to differentiate the two.

Also added a conditional to determine the color of the pin.