fix(frontend): Make pill name re-expand to full-width if only one
There was a small ternary in place to switch the max-width between `15ch` and `undefined` based on whether or not there was more than one feature selected.

However, setting the CSS rule to `undefined` does not remove the previous value. A quick look at the docs for `maxWidth` shows that `none` should be used to negate a value.