Build times (#1237)
* fix(devops): Run workflow steps in parallel

Instead of running steps more linearly, I have tried to run the tests
and build the docker images in parallel. This doesn't reduce the amount
of minutes we use, but it does run them in parallel and will speed up
the workflow as a whole.

* fix(devops): Fix caching of node_modules

It appears we have had a broken caching of node_modules for some time
so that the frontend packages are always re-installed. I have now
updated the workflow to correctly cache just the `node_modules`,
check if they are outdated and if so, re-install, otherwise, use the
cached version and skip the install.

* fix(devops): Update secret to use organisation secret

Beforehand this workflow was using a repo-specific secret. I have
now updated it to use the organisation secret as it doesn't change
between repos.

* fix(devops): Make caching more specific

I have made the caching of client dependencies far more specific,
in case we have `yarn.lock` and `node_modules` in non client related
directories.

* fix(devops): separate out dependency installation

To try to speed up builds, I have separated out the frontend dependencies
installation, so this will exist as a separate layer, we can cache.
Then we should be able to use this layer as a base for future builds.

Co-authored-by: Mark Small <mark.small@astrosat.space>