feat(backend): refactored backend checks for the size of the AOI
Added validation checks on the SatelliteSearchModel &
SatelliteSearchSerializer to ensure that the AOI is smaller than the
maximum size.

This was a bit complex b/c in order to compute the area I had to project
the AOI from a lat/lon geometry onto a flat plane geography; `GeoPandas`
helped select an appropriate CRS to use w/ this.