chore(celery): Reduce frequency of celery pings (#1307)
* chore(celery): Reduce frequency of celery pings

It turns out these celery pings used for healthchecking the celery container are quite expensive to run, each check
using 100% CPU for multiple seconds. This is chewing up a significant chunk of our EKS Workers' CPU, quickly runs our
CPU Credit Balance to zero, and impacts the performance of our other services.
This commit reduces the frequency of these celery healthcheck pings to save resources. It's safe to do, because celery
is only used for non-urgent background tasks, so it doesn't matter if they happen after several minutes, so we can
safely increase the interval between healthchecks (and possible restarts) to several minutes.

* chore(celery): Fix other healthcheck parameters

To suit the reduction in healthcheck frequency