fix(frontend): Specify the name of terms doc to use
The backend is failing to return a document as there can be multiple
documents of types. I have narrowed the request so that only active
terms are queried and for the document displayed in the profile
component, I have named the actual document to return, so it only
finds one.

IssueID ORB-902