feat(backend): Add ability to post JSON in PRoxy requests
The existing code threw errors when trying to POST JSON data in the
request to the PLD API. I have therefore added code to check the
method type and if POST, I assume you want to post JSON data. This
may not always be true, not every API will work with JSON data, but
this works for what we need right now.

IssueID ORB-874