feat(backend): Simplify Document constraints
Instead of using a Nullable BooleanField, I've added extra fields
to the constraints meta field. This allows us to have many is_active=False
but only one is_active=True. There is a migration for existing data
and a schema migration to deal with the model change. It is possible
to still generate a dbase IntegrityError, so added a ModelForm to
check the data before trying to save it to the dbase.