fix(frontend): Add color override for each checkbox
For this task, I needed to extend the re-usable CheckboxFilters
component, so that it can take a color override, for each individual
checkbox. Before, the background color was always the same, or used
a color map, which dictated the colors. For this, we need to be able
to specify the color of each checkbox used.

IssueID ORB-851