feat(backend): added the ability to proxy requests using socks5 (#1394)
* feat(backend): added the ability to proxy requests using socks5

In order to work around gaining access to data-sources-directory during
development when the VPN is down we can either replace the URL to that
app and the associated data token secret w/ the STAGING rather than
TESTING values.  Or we can setup a proxy and add those details to
"server/.env.local".  This PR enables the latter.