feat(frontend): Add drawing of general AOI
Renamed satellite code relating to drawing of AOI to be more explicit
in that it is for satellites, so it doesn't clash, with the more
generic general AOI drawing code.

Added code, based on satellite drawing of an AOI, that will be the
basis of what is saved and used in dashboards.

IssueID ORB-720