fix(frontend): Set default color for pin and group clicks
If there are multiple features at the same location, then a group
icon is used. I've set a default colour for these as they could be
different status', so can't use one of the existing colours.

Added the ability to click on a pin that has multiple features, when
this is the case, both features are displayed in a popup, when clicked.

IssueID ORB-980