fix(ir-proxy): TEMPORARY Disable checking SSL Certificates (#1442)
The NS IR API currently uses a self-signed SSL certificate, which means
it doesn't pass SSL verification. In order to use the API right now we
have to disable SSL verification for the proxy requests.

This change MUST be reverted as soon as the NS IR API supports real SSL
certificates again.