fix(frontend): Limit AOI to just ORBs that are permitted
I've made it so the AOI tab is only visible to users with a licence
to the `Eco an Alba` orb.

I also updated the icon for the tab as it has now been provided by
the users. The icon was added to the UI library, so that is why I
updated the version of it.

IssueID ORB-720