fix(frontend): Add ability to edit an AOI
The primary focus of this commit was to add the ability to edit
an AOI.

I did add some other stuff though that I did and was hard to then
separate out for the commit. like getting a screenshot of the map
to save with the AOI. Passing in functions to manage the update/deleting
of an AOI etc.

IssueID ORB-720