fix(frontend): Add fetching of data from IR
I added the `useEffect` to fetch data via the PROXY, from the IR API.
I then take that data and update some local state, which in turn is
used as props to the sub-components.

I also played around with the styling to make it look a bit more like
the design.