fix(frontend): Update to NodeJS 16.13.1 and fix any bugs
I've updated the project to ensure the frontend uses NodeJS `16.13.1`
by setting it as the value in `.node-version`. I also updated a
number of packages, as I had errors. `babel-loader` should never have
been installed on it's own as it is provided by CRA, but another
package was wanting a different version and CRA complained. The best
solution seems to be to set the `resolutions` object in the package.json.
I also did this for `react-router-dom`, even though I removed any
wildcards relating to it, `connected-react-router` and `history`
packages.