fix(frontend): Refactor Progression vs Planning chart
I created a brand new component instead of the more inline way it
was added to the WalthamForestConfig component previously. I removed
the usage of StackedBarChart component and just replicated it since
that is already wrapped in a BaseChart. The purpose of that was so
we could pass in a Legend component to represent this individual
chart.