feat(frontend): Upgrade to React Router 6
This is a highly invasive change, means upgrading react itself, then
React Router as well as other dependencies, react relies on. That in
turn meant I had to change some linting rules to be less strict, before
I couldn even get the app to start.

I then proceeded to upgrade the `routes` we use for internal routing
to be all public, before finding a new way to make some private and
correctly redirect, when not logged in.

I then updated the rest of the routes in the app, following the same
concepts.