chore(backend): fixed intermittent failing test
The LicenceFactory used in tests randomly assigns an access level.  The
generate_token fn uses licences to identify DataScopes to buid a JWT
from.  If a licence happens to have no access then it won't be included
in that JWT which made the test fail.  I simply hard-coded the test to
always give the licence "READ" access.