feat(frontend): Use MSW for tests
I have removed `jest-fetch-mock` in favour of using the newer MSW
library. This does a very similar thing, but instead of using multiple
dependencies to achieve it, we just need one. MSW will also be used
for our dev backend, so we can work on the frontend while the backend
is under development.