fix(frontend): Updated code due to review comments
I've removed a few `console.log` statements and some commented out
code, I missed when creating the PR.

The PNG file I suspect is due to a test, or the dev proxy server
trying to save the file. We may need to add a `.gitignore` rule to
stop this, not sure what is the best answer right now, so I just
removed the file itself.