fix(frontend): Fix clicking disabled buttons
This is a bit of a weird one, after upgrading some dependencies, some
tests that clicked on disabled buttons to test they didn't fire actions,
started to complain about inline CSS in parent object and pointer-events.
This solution basically skips the check, not ideal I agree, but I couldn't
find a better way to prevent the test from failing.