fix(frontend): Fix failing test
I'm not sure why this was working, but after upgrade of libs, it broke,
I'm not exactly sure what lib upgrade broke it as I did a number at once.
However, having read into jest timers and watched Kent Dodds, I feel this
is the right approach, at least it seems to be what others use.