fix(frontend): Update navigation URLs
Having changed the routing library, I have had to update the links
as they cannot exactly match what is in the constants file. The routing
is more parent/child now, that before, so the constants work for
the routing to components, but links need to include the parent and
child parts, hence the addition of `/accounts` to these navigations.