fix(frontend): Add routes for updated Mission Control tests
The mission control code is more modular, no longer using mission-control
and store as the brains, passing in props to the components. This was largely
forced upon us by the new style of react-router. The outcome is that the tests
now have to be wrapped in `Routes`, so that when the component switches from
route to route, there is a valid route for the test to route to.