fix(frontend): Update form to latest react-hooks-form method
The react-hooks-form has changed how you are meant to use it, so I
updated the component to comply and then the test that broke.

I don't know why, but I've had to wrap each `TextField` in a `Controller`
component from the library, yet I don't have to do this in other forms.