fix(frontend): Improve URLs used accessing accounts views
We used to be able to re-use the same constants for `routes` and
`links` but the upgrade of react-router` now works more on parent/child
relationships, so this is no longer possible. Therefore, I have some
constants for `routes` and then new ones for the URLs, which still
need to be absolute.