fix(frontend): Fix resending of password reset email
The initial call to send a password reset email works fine, however,
if you lose that, there is the ability to resend the email, however,
that was failing. It turns out the resend was just passing the email
address, whereas it should be wrapped in an object. Doing so in the
password reset success view fixed the problem.