fix(frontend): Fix problem verifying email after registration
Once a user has registered, an email is sent so they can verify it.
The functionality to verify this was not being triggered, due to the
`match` params no longer being passed to components of a route. Instead
I use the `useParams` hook to get them and now the code works again.
Since it uses `Route` params, I also needed to wrap the component in
the test in a `Route`, so it could get the param.