fix(backend): update which licences are returned from assign_licences
During assign_licences, after assigning a licence to the specified orb
there is code which removes any existing licences that are exclusive.
An error was ocurring if a licence that had been removed was still
returned from assign_licences.  This PR only returns licences that have
not been removed.

IssueID #ORB-1108