fix(frontend): Revert deck.gl updates, update remaining packages
This removes the carats from the deck.gl packages, pinning them to the versions known to work. They are still slightly outdates, but not by a major, only minors.

The others, with the exception of `react-map-gl` have also been upgraded.

The map, layers, analysis toolbar are stable as of this commit.