fix(frontend): Fix error during bookmark deletion
I mistakenly read the response from the API request to delete a
bookmark, forgetting that this doesn't return anything. I have now
removed the code to read the json from the response and return it,
as well as update the corresponding test, that started to fail. I've
not tested it and ran the app and successfully deleted bookmarks.