fix(frontend): Fix reference to `ref` warnings
We were passing references to `ref` down to the textfield and it
was complaining in tests. I therefore extract the ref from the
field and pass the rest to the textfield, which fixes the warning.

Tests complained about setting the value to an empty string in the
dropdown. I therefore replaced the constant value to use `NONE`
instead. We also used the value as the key, which was failing since
it was empty. I therefore changed it to use the name, not the value
so it was a bit more unique.