fix(frontend): Convert feature properties from strings to complex objects
When the data is loaded, complex feature properties such as objects and arrays
are stringified. I've added code to update the string representations to
whatever they should be once when rendering. We want this to be done in one
place as there are a number of other places where the feature properties are
used and we don't want to parse them multiple times.