fix(frontend): Repurpose slice function, integrate functionality into dashboard
This updates the slice function that was previously for updating only target values, to allow updating of the user's orbState for any particular dashboard, with any data.

Then, a function was created to dispatch an object parameter along with the dashboard's source_id and the user's user object, and that function was passed to the various components that need to persist their state.