fix(frontend): Update logic, fix type
Updated the form logic to set the targets to an empty object in the dashboard, earlier than when initializing the state in the form. the reason for this being that we want to compare the state in the form to the targets from the API to determine if the form has been changed, but by having the OR when initializing state, the API targets will be converted from `undefined` to an empty object, and the form will read that as now being different, and will disable the button.

Also changed the target type from an array to an object.