fix(frontend): Make tooltip more generic
Moved tooltip content to generic components directory, so it is
available to all features, not just iso+ radio buttons.

I then added the tooltip button to the checkbox filters, only if
there is an `info` prop (comes from metadata).

I then widened the list, so it didn't look too squashed.