feat(frontend): Add Nature Scotland Protected Areas layer
Added new config specific to the Nature Scotland Protected Areas
data. This layer is an MVTLayer, similar to iso+, but uses data
filtering and a patterned style, instead of just the normal fill,
we have done up to this point. It has a click handler, so that
feature info is displayed in a popup, if clicked.