fix(frontend): Enable allowing of missing API/target data, refactor grouped bar chart
This allows wither the API or target data to have missing values, or to have date
ranges that extent further into the past/future than the other.

This is done by using both datasets to create a combined "timeline" of the earliest
and latest years, then padding any years with no data with `null` values.

I also refactored the TotalHousingMultiChart to contain all of its own logic and functions,
instead of having previously processed data passed in.

The goal of this is to have a more encapsulated component, seeing as the processing
functions have become very specific, and easier to test by passing data in the API's
non-processed shape.