fix(frontend): Update logic, add tests, update dialog screens and constants
This updates more of the logic to match what is desired, after a meeting with BAs.

The targets dialog has been updated to show the correct number of years for datasets,
dataset names updated, and a function call to automatically render from the current
year and back.

Also added tests, and removed unneeded constants from the WFC constants file, which
were the hardcoded input field placeholders.