fix(frontend): Improve styling of filtering components
I reduced the amount of padding at the bottom of the heading as it
seemed too far away, from the dropdown. I also separated the dropdown
and the DateRange filter into separate grids, again, better spacing
and nicer on the eye. Lastly, I reduced the padding on the input
box for the DateRange filter's input field, as you couldn't read the
whole date.

IssueID ORB-1013