fix(frontend): Fix pin colours used
I previously changed these to match status values, basing my decision
on what I was seeing in the data. I made a wrong assumption though,
so after discussion, I've changed the colours to be key'd off of the
`Status Category` property of the feature, instead of the `Status`.
Doing this means I needed to update the list of keys in the `PIN_COLORS`
object to match.

IssueID ORB-1013