fix(frontend): Add targets to dashboard settings, use local state
Added the targets to the dashboard settings along with the dashboard's toggles/dropdowns.

Added some local state for the targets that is initialized to the targets fetched from the server,
then manipulated at will by the user with no network requests.

The entire dashboard settings object is then dispatched to the server when the user
either closes/refreshes the tab or navigates away from the dashboard.