fix(frontend): Split user orbState into targets and settings
This is a big, quite invasive change, but it is also necessary and it's
better to get it out of the way now (day before new sprint starts).

Before, the targets (objects: { 'yyyy-yyyy': 123  }), and the settings
(strings: 'Socially Rented') were at the same level in the user orbState
object. This causes problems when it comes to runing the target processing
functions on the target objects, as not all of them are objects.

By separating them out, there is a clear distinction between the two.
This also means that it is much easier to see what is going where,
instead of having to log out generic-sounding variables like 'orbState'.