fix(frontend): Changed structure to accept new API data shape
Bit of a big one, but it's actually just lots of small changes.

Basically, the shape of the live API data is different from the mock data
used previously, so a lot of stuff had to be changed to make it work.

Also, the values used for the data has been changed to straight integers such as
`2018` instead of the hyphenated equivalent such as `2018-2019`. This meant
further changes, and also a refactoring of the dialog screens to make sure
target data matches, which broke pretty much all of the tests involved with
this dashboard.