fix(frontend): Make date type for filtering configurable
I have made the date type used to filter data configurable from
metadata. I have done this as we are supposed to use `Decision Date`
but it doesn't exist in the data, so I have used `Commencement Date`.
I suspect `Decision Date` to be added to the data, so when it does,
it is just a config change, not a code change, needed.

IssueID ORB-1188