fix(frontend): Refactor charts top new structure, adjust styles
This basically restructures how the charts are shared around the codebase.
By allowing them to be rendered as children, it eliminates the need for
renderProps functions to be passed, and also many of the JSX-related bugs
we'd been seeing recently.

This also opens the door to re-usable charts down the line, but there isn't
time for that at moment.

Also adjusted some styles to match the new structure, and designs better.