fix(frontend): Refactor to use objects for legend data
Refactored the two housing delivery charts to use an object for reference
when looking up colors, as this ensures the legends match the data.

Using indexes to pull colors out of arrays is not a guarantee that
the colors in the legends match the data, as arrays cannot guarantee
the same indexes every time.