fix(frontend): Refactor date range filtering
I've refactored how the date filtering works, with regards how to
calculate if a feature is within a date range. I have removed the
default range, which is no longer useful.

I also refactor how the sidebar deals with getting the latest date
in the features, to calculate the default date range.

IssueID ORB-1188