fix(frontend): Fix warnings in tests
When we run tests, we see a lot of warnings, some are due to bugs
in the UI library (which are being fixed also) and some are bugs
in orbis code. It is these orbis code bugs I have fixed here, or
at least some of them. These are the fairly easy ones to fix, others
I fixed, but broke the application, so had to undo.