feat(frontend): Add date range to form
Created a wrapper component for any form element needing a fieldset.
I updated the existing description to use the field wrapper and added
the date range component to update the form with the date range selected.

I also added the hidden panel bit, so that when ready, the activities
lists can be placed inside this hidden area and shown, when button
clicked.

IssueID ORB-1164