fix(frontend): Refactor stack total function, refactor affordable housing data percentages
1. I realised that the getStackDatumTotal could be simplified, as one of the arguments was
not necessary, so refactored where it appeared in the app, and updated the tests.

2. Added a function that filters all data elements from the Affordable Housing that do not
have a "twin" in the other dataset, as only pairs can be displayed on the chart, making these
lone data elements useless. I then added tests to make sure it does as expedcted, and  passed the filtered datasets to the timeline function, to be spread onto the chart.