fix(backend): Fix bug calculating distance
Since the projection of the AOI and BBOX are in lat/lon (EPSG:4326),
which uses degrees, distance calculations we currently have don't
work correctly. I have now re-projected the AOI and BBOX to easting/
northing (EPSG:27700), this uses meters. Therefore, the distance
calculation is easy to do to get the distance in kilometres.

IssueID ORB-1247