fix(frontend): Date validation
We were setting the minimum date config to today, Yet setting the
date to today in the app, showed an error, stating it can't be before.
The reason is the filter was set to midnight on the date and the
value when selected/entered was the datetime, which is afterwards.
I therefore set the today date in the generic widget to only set
the year/month/day, not the time aspect. This should fix the problem
without affecting other uses of this widget.

I set the today date to the string `today`, not a string representation
of today's date.

IssueID ORB-1124