feat(backend): Add proxy for IR typeahead Search
This adapter is specific to the typeahead searching, where the response
is just an array of `name` and maybe `id`. So the `name` is displayed in
the dropdown, for the user to select, when they type into the typeahead
field. The `id` is used for when they have selected aname from the list,
it is used to search for that specific suggestion.

IssueID ORB-1123