feat(frontend): Add typeahead UI
I've added the frontend code related to a typeahead. There are 2 hooks
`useDebounce` and `useThrottle`. Each of these are useful under different
circumstances. The typeahead component is just a textfield and list to
display the results. The AOI component uses the typeahead and has a
hard-coded URL for the moment, this is more of a work in progress
at this time.

IssueID ORB-1123