feat(frontend): Manage for submission and passing of results
I have added the code now to make the API request, when the user
submits the form and extracts the response data to pass to the
individual components that will display the data.

On submission, the UI also switches to the Results tab, so the user
can see the results.

IssueID ORB-1261