fix(frontend): Remove form from date range filter
The date range filter wrapped a form element. While it does use the
react-hooks-form library, it doesn't need to be an actual form. This
component is widely used in other forms and you cannot have a form
inside another form. I therefore changed this component to be a `div`
instead. We can still invoke the onchange event handler on this, so
there is no change in functionality.