feat(frontend): Add form handling to fetch activities
This code handles what is needed to fetch the activities list from
the backend. This replaces the hard-coded activities list we currently
use. I have also removed all the list reduction code as that was
never going to stay. This means there are many more options in the
shuttle list, but that is fine.

We still have to better validate the form e.g. min date is today.

IssueID ORB-1261