fix(frontend): Add id to each assessment shuttle texbox
For some reason, when you click on the textbox over the placeholder/label
it doesn't move the placeholder to be a label, but does when you click
to the side of the placeholder. Adding an `id` to each textfield seems
to fix this. I repeated this with mui's own example and not having
an `id` made no difference. It's confusing, but not work the effort
spending more time on.

IssueID ORB-1264