fix(backend): Add extra fields and change activities
I added a couple of new fields, to save the associated geometry and
report generation date.

I changed the list of activities associated with a proposal to be a
list of dictionary objects. I then had to update the serializer
validation function looking for duplicate activities as well as
tests.

IssueID ORB-1263