fix(frontend): Fix Impact assessment form layout
I previously introduced a bug when adding the Yes/No dialog, for
when a user closes the Impact Assessment form. I added some CSS
that made it so you couldn't see the whole form, this was primarily
caused by the `justifyContent: 'flex-end'` code, which I have now
removed. I have instead added this to where it should be, inside
the Yes/No dialog itself.

IssueID ORB-1264