fix(frontend): Remove new state in favor of existing
I create a new bit of state to hold the current state of the form.
It turns out that was unnecessary as the `initialFormState` is updated
to reflect the currrent state. I have removed the new state and
renamed the existing `initialFormState` to `formState`, so it always
reflects the forms state, not just at the beginning.

IssueID ORB-1163