fix(frontend): Space chart wrapped items better
When there isn't much content, ChartWrapper children are placed at
the bottom due to flex rules to `space-between`. So, if you only have
a single line of text and the hight of the wrapper is large, the text
is way down. I've changed the justification of the wrapper to use
`flex-start` to avoid this. If there is little in the wrapper it is
top-aligned and will grow as required.

IssueID ORB-1286