fix(frontend): Display message when no protected areas data
There can be the situation where there is no `protected areas` data.
In this case, we need to inform users, this is the case. I also
amended the text telling the user about the AOI, as it stated `overlap
or intersects`, this is not true, it is only if it `overlaps`, so
amended the text accordingly.

IssueID ORB-1286