feat(backend): Get map styles from server
The map styles are hard-coded, but now we have the ability to get
them from the server-side. This makes it much more configurable, in
that they are aligned to a customer, who can decide which to show
to their users. Right now, the frontend is still using the old
setup, this is just setting the ground work to switch to making them
more configurable.

IssueID ORB-945