fix(frontend): Use map styles from API
Now that we get map styles when a user has logged in, from the API,
I put those into redux state, rather than hard-code them. This also
involved some minor tweaks to the switcher, to use new data-shape and
updating the function that decides whether to transform the data
requests, or not.

IssueID ORB-945