fix(backend): Update backend to extract newest data for contact details
The API has now been updated, to return the nearest Nature Scotland
office details, to the AOI searched for. We used to get this from
each suggestion, then get rid of duplications, but now, it is a top-level
key/value pair. I've moved the code to extract the data out of the
for loop, to just above it and added a new `area_office` entry for
the contact details.

IssueID ORB-1057