fix(frontend): Fix map style tests when logging in
I forgot to check the tests when fixing the initial bug, which I
have now done so. We peer into the user response, but since the
test had nothing for it, it was failing. I now act more defensively
when checking of user data exists, to then put it into redux, thus
fixing the tests.