fix(frontend): Fix failing network request for tascomi dashboard
Since we don't yet have the right data, we just implemented the idea
of fetching some data, even if it isn't the data we really want.
That gets more of the work to manage the fetching of data out of the
way and we just need to plug in the right URL.

However, the URL being used consistently failed, so, since it was
an easy fix, I just went ahead and did it.