fix(frontend): Remove unused colour settings
Noticed during Code Review, but I accidentally left the `pattern`
config, when it isn't used. I used a previous config as the basis
for this and forgot to remove these attributes.

I also fixed comments that were specific to the dataset config I
used as the basis for this one.

IssueID ORB-1115