fix(frontend): Fix typed text not visible in textfield
Setting the value with the `||`, to fix the uncontrolled component
to a controlled one, broke the actual working, while fixing the test.
I have now updated it so that the text is visible when typed on the
field and when an option from the dropdown has been selected.

I also removed an import that was auto-inserted by the IDE that isn't
being used.

IssueID ORB-1026