fix(frontend): Fix PLD Data filtering
The filtering was including features that did not have the chosen
date type e.g. `decision_date`, `actual_commencement_date` etc. I
have put a check in as the first thing in the filter to check that
the date type exists on the feature, or reject it and that has fixed
the problem.

IssueID ORB-1025