fix(frontend): Update name displayed for Grid refs
I've noticed that if you search for a grid ref and get a `protected-area`
type, it doesn't have a code, yet if you search for a place and get
a `protected-area`, it does. The result was sometimes I was seeing
`undefined` in the list item text, which is bad. I now check that
the `code` exists and only then add it, which does mean we can have
multiple list items, with the same text.

IssueID ORB-1026