fix(frontend): Set transparent opacity
If there is no data in the metadata, indicating there timestamp doesn't
exist. Set the polygon's opacity to be zero i.e. clear. This means the
color won't be visible. I also make the onHover and onClick functions
not work in this situation, so you can click where there is not timestamp
data.