fix(frontend): Improve resetting global state on logout
Before, when we logged in, I forced fetching of sources, but then
the code already in the landing page was obsolete. I think this is
a better approach, on logout, we reset the relevant state. This then
ensures the code run on first login, again runs, when you login again
as a new user.