fix(frontend): Extract logic from progress indicator, move up to usage component
This moves the logic that decides the amount of "coloured bar" to show on the progress wheel
itself, as well as the component(s) to display inside the indicator, to outside the component
and up to the component where the indicators are used.

The reason for this is that the indicator is now required to be used for much more generic
purposes besides just those used for WFC.

This way, each usage of the indicators can have its own custom logic/display component(s),
without the need for mass refactoring and heavy testing.