fix(frontend): Update styling of info when unclicked
The info button, when toggled off, was not returning to it's original
colour. It does when you click away from it, but not re-click it to
toggle it. I have now got this working, but with some changes, namely
the styling of the unclicked button should follow the theme and not
be overridden, this causes a lot of problems remembering the new
color.

IssueID ORB-866