fix(frontend): Fix warnings displayed for client code
There are too many warnings that we should be dealing with, being
displayed in the console. Due to this, devs are starting to ignore
them. This is bad practice, so I have fixed what warnings I can, so
that future warnings do register with devs, so they can fix them.