fix(frontend): Get indicators to display from config
Having the list of indicator types hard-coded is asking for future
problems. Better it is passed into the dashboard, so we can change
it without making code releases etc.

I also took the opportunity to rename a few variable to try make
the naming more consistent.

Another thing I noticed was, we were trying to display an indicator
for every entry in the API response. It could be we don't want that
though and only want to display some. Therefore I refactored the
code to only extract what is in the metadata, not try to find metadata
for each data entry in the API response.